Fix issue when searching with instance counts
- When more than one instance of a letter appeared in a word but
one of the letters was marked as not existing in the word, the app
was not removing the letter from the "Not Exists" collection after
determining the instance count. This caused the search string to be
invalid by searching for a word that both did and didn't include a
letter.