Add frontend application (#1)
* rename frontend to frontend-sample

Want to build the UI from scratch rather than
use something existing from an old app

* Rename Dictionary.Data to Dictionary.Business

I felt that "Data" was an improper namespace name. Renamed
to "Business" to more appropriately match the functional
purpose of the project

* Initial commit frontend

- All boilerplate code, nothing substantial yet.
- Using NextJS as the framework with NextUI as the component library

* Add basic word tiles

- Still need to add interactability to each tile.

* Include theming

* Add onscreen keyboard

- Add ability to change word status

* Add word puzzle board object

- Also add a logic class to handle the state changes of the board
for any inputs

* Hook up on screen keyboard

- On screen keyboard can now edit words

* Change board style

- Alter the colors of the board to make things look a little more
presentable

* Delete word when backspacing with no letters

* Add positional searches

- Allow to exclude or include letters at specific positions of the
word

* Fix tests

* Add rewrite rules to Next config

- The API runs on port 5000 in development, while the Next app runs
on port 3000. A rewrite rule needed to be added so that any API calls
from the client are correctly redirected to the API server. In
production, this is solved through the use of a reverse proxy.

* Add first API call

- No real UI implemented with the results, but added an API call for
the word search. Still need to properly build the query from the
word board.

* Include use-case in tests

* Convert board to word search query

* Refactor puzzle breakdown logic

* Show search results

* Improve base styles

- Added flex styling to main board page to improve spacing of board
and keyboard
- Added loading properties to search button to indicate when
search is in progress
- styled search button

* Fix scrolling issue with search results in modal

- It's a hack, since it seems like there is an underlying issue
with the Table component in NextUI. It doesn't scroll properly and
the height property is ignored if passed in with CSS. Added a wrapper
div that instead scrolls if the table is too long.

* Fix page size on mobile

- 100vh does not work properly on mobile, because it does not consider
the address bar of the browser which can pop in and out and change the
available screen space.

* Add media styles for keyboard

* Fix words being cutoff on small screens

* use variable for screen size

* Fix warning about nonexistent dom element

* Remove unused div object

* Fix arbitrary button padding on iphone

* Add ripple effect to keyboard

* Accept input from physical keyboard

* Remove log statement

* Fix event listener being called multiple times

- The react effect was being called many times (once on each render).
Added a dependency array to avoid multiple calls to the effect, but it
was still being called twice on first render. Added a variable to check
and prevent the second call from occuring.

* Make word in search results uppercase

- In order to match the case of the word game, I felt it made sense
to make the words all uppercase.

* Add validation checking to word puzzle board

* Add filtering of letter counts in word

- For example, if two "B"s exist on the word, then the application now
checks for that possibility depending on the words given in the guesses.

* Add unit tests

* Improve board positioning and styling

* Add rudimentary navbar

* Remove console warning

* Add Responsive Modal component

- Shrink modal on larger screens

* Delete sample project

- Remove project used as a reference on initial setup of project

* Add error handling

* Add error handling to frontend

- Added logic to handle API errors and to sanitize the responses
received from the API.

* Add validation check when no results were found

* Add help instructions modal

- Add help information that walks a user through how the Searchle
application works.

* Add word count validation

- Don't allow more than 10 words per search. This is an arbitrary limit
to prevent users from going crazy and adding hundreds of words to
cheat on a word game that allows only a max of 6 words.

* Add ability to clear the board

- Added a clear button next to the search button. A user is now able
to clear the contents of the puzzle and start again without reloading
the page (As was previously necessary)

* Delete old sample css files

* Remove larger buttons on bigger screens

- I ended up not liking the look of the bigger buttons, so I'm reverting
those changes.

* Disable text selection on board letters

* Remove focus outline on board

* Clear validation messages on board reset

* Fix spelling mistake

* Remove "Unresolved" as a selectable status

- Had a friend test out the app, and their main feedback was that they
did not like how the letters didn't start as black. They felt they
were being asked to click unnecessarily

* Fix issue when searching with instance counts

- When more than one instance of a letter appeared in a word but
one of the letters was marked as not existing in the word, the app
was not removing the letter from the "Not Exists" collection after
determining the instance count. This caused the search string to be
invalid by searching for a word that both did and didn't include a
letter.